+++ v2/magic.js
@@ -1,4 +1,4 @@
-import { view } from './mem.js';
+const { view } = require('./mem.js');
 
 const formats = {
   ttf: { type: 'font', format: 'ttf' },
@@ -10,7 +10,7 @@ const formats = {
   tiff: { type: 'image', format: 'tiff' },
 };
 
-export function buffer(init) {
+function buffer(init) {
   const u8 = view(init);
   if (0 === u8.length) return;
   if (0xff === u8[0] && 0xd8 === u8[1] && 0xff === u8[2]) return formats.jpeg;
@@ -21,3 +21,5 @@ export function buffer(init) {
   if (0x47 === u8[0] && 0x49 === u8[1] && 0x46 === u8[2] && 0x38 === u8[3] && 0x61 === u8[5] && (0x37 === u8[4] || 0x39 === u8[4])) return formats.gif;
   if (0x89 === u8[0] && 0x50 === u8[1] && 0x4e === u8[2] && 0x47 === u8[3] && 0x0d === u8[4] && 0x0a === u8[5] && 0x1a === u8[6] && 0x0a === u8[7]) return formats.png;
 }
+
+module.exports = { buffer };
